requirements.txt
numpy~=1.23.5
PySide6~=6.4.2
opencv-python~=4.8
PyWavelets
Pillow
reedsolo~=1.7.0